requirements.txt
scikit-learn
numpy
opencv-python
Pillow
joblib
streamlit
